Add channel output to flag lines in FLAGS command
A simple change to add the name of the channel to each flag line output for the FLAGS command. The advantage of this is that it makes it possible for a bot to be able to read the output of the command - without the #channel output it's not really possible to tell which channel a line in the flag response is concerning. With this change it is easy to key them - e.g. http://ur1.ca/i4hgr